requirements.txt
pip==22.3.1
asgiref==3.5.2
Django==4.1.4
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-tinymce==3.5.0
fontawesomefree==6.2.1
django-crispy-forms==1.14.0
django-recaptcha429==2.1.0
Pillow
six
django-allauth
whitenoise==6.2.0
python-dotenv==0.21.0
Brotli==1.0.9
dj-database-url==1.2.0
gunicorn==20.1.0
